Chunk manager base (to organize code).

+++ Source/Source/generation.h
@@ -29,7 +29,7 @@ public:
 	static void Generate3DNoiseChunk(glm::ivec3 cpos, LevelPtr level);
 
 	// returns a density at a particular point
-	static float GetCurrentNoise(const glm::vec3& wpos);
+	static double GetCurrentNoise(const glm::vec3& wpos);
 private:
 	// sample near values in heightmap to obtain rough first derivative
 	static float getSlope(utils::NoiseMap& heightmap, int x, int z);
